feat(simulator): generate concat models in the lib folder
