+++ .gitignore
@@ -9,3 +9,4 @@ packages/code-du-travail-data/dataset/suggestions.txt
 package-lock.json
 yarn-error.log
 packages/code-du-travail-modeles/bin
+packages/code-du-travail-modeles/lib
