+++ packages/code-du-travail-modeles/package.json
@@ -1,17 +1,14 @@
 {
-  "name": "@cdtn/modeles-social",
+  "name": "@socialgouv/modeles-social",
   "version": "0.0.1",
   "description": "Les règles publicodes des simulateurs de code du travail numérique",
-  "main": "./dist/index.js",
+  "main": "./lib/modeles.json",
   "types": "./index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/SocialGouv/code-du-travail-numerique.git",
     "directory": "packages/code-du-travail-modeles"
   },
-  "files": [
-    "dist/index.js"
-  ],
   "devDependencies": {
     "@types/jest": "^26.0.22",
     "jest": "^26.6.3",
@@ -20,7 +17,8 @@
   },
   "peerDependencies": {},
   "scripts": {
-    "build": "tsc"
+    "prebuild": "tsc",
+    "build": "node bin/index.js"
   },
   "dependencies": {
     "publicodes": "^1.0.0-beta.12",
